// src/utils/types/ReportAnalysisItemType.ts
export type ReportAggregateItemType = {
    name: string
    percentageLoanValueChange: string
    percentageLoanVolumeChange: string
    percentageTransactionValueChange: string
    percentageTransactionVolumeChange: string
    totalLoanValue: number
    totalLoanVolume: number
    totalTransactionValue: number
    totalTransactionVolume: number
}